+++ tests/testthat/test-filters.R
@@ -37,3 +37,23 @@ test_that("filters return values that are not matched by the rules", {
   fails2 <- filter_fails(rules, data, per_rule = TRUE)
   expect_equal(fails2, fails)
 })
+
+test_that("No fails do not crash", {
+  data <- mtcars
+
+  rules <- ruleset(
+    rule(mpg > 10 & mpg < 35),
+    rule(cyl != 5),
+    rule(vs %in% c(0, 1))
+  )
+
+  res <- check_data(data, rules)
+  rr <- filter_fails(res, data)
+  expect_equal(nrow(rr), 0)
+  expect_equal(names(rr), names(data))
+
+  # also works when ruleset is provided
+  rr2 <- filter_fails(rules, data)
+  expect_equal(nrow(rr2), 0)
+  expect_equal(names(rr2), names(data))
+})