+++ NEWS.md
@@ -1,7 +1,8 @@
 # dataverifyr 0.1.6
 
+* allow multiline rules in yaml file
+* fix bug where no fails in filter_fails would result in error
 * Fix bug where multiline rules would break
-
 * fix minor error in Readme
 * `filter_fails()` allows the first argument to be a `ruleset` and not only a result of `check_data()`